refactor(site/src/pages/AgentsPage): extract shared ChatMessageInput story helpers
Add a co-located storyHelpers module with makeUserSkill, shared
mockSkills, and the visible-text query helpers, removing the duplicate
copies from the ChatMessageInput and PersonalSkillsTriggerMenu stories.
